add: 최근 본 상품 추가(훅 추가)
