+++ frontend/client/src/pages/GoodsDetail/GoodsDetailPage.tsx
@@ -8,11 +8,14 @@ import GoodsImageSection from './GoodsImageSection/GoodsImageSection';
 import RelationSection from './RelationSection/RelationSection';
 import { getGoodsDetail } from '@src/apis/goodsAPI';
 import { usePushToast } from '@src/lib/ToastProvider/ToastProvider';
+import useRecentGoodsHistory from '@src/hooks/useRecentGoodsHistory';
+
 import theme from '@src/theme/theme';
 
 const ERROR_SERVER = '서버 문제로 상품 정보 조회에 실패하였습니다!';
 
 const GoodsDetailPage = () => {
+  const [recentGoodsList, setRecentGoodsList] = useRecentGoodsHistory();
   const { id } = useParams();
   const pushToast = usePushToast();
   const [goods, setGoods] = useState<DetailGoods | null>(null);
@@ -21,6 +24,7 @@ const GoodsDetailPage = () => {
     try {
       const data = await getGoodsDetail(goodsId);
       setGoods(data.result);
+      setRecentGoodsList([data.result, ...recentGoodsList]);
     } catch (e) {
       console.error(e);
       setGoods(null);
