+++ frontend/client/src/pages/Main/Main.tsx
@@ -5,29 +5,12 @@ import GoodsSection from '@src/components/GoodsSection/GoodsSection';
 import PromotionCarousel from '@src/components/PromotionCarousel/PromotionCarousel';
 import { Promotion } from '@src/types/Promotion';
 import Footer from '@src/components/Footer/Footer';
-import SideBar from './SideBar/SideBar';
+import RecentGoodsView from './SideBar/RecentGoodsView';
 import { getMainGoodsListMap } from '@src/apis/goodsAPI';
 import { getPromotions } from '@src/apis/promotionAPI';
 import { useRecoilValue } from 'recoil';
 import { userState } from '@src/recoil/userState';
-
-const mockProductImagePath =
-  'https://user-images.githubusercontent.com/20085849/128866958-900ad32a-cd32-4b97-be79-1dbbc9dcb02d.jpeg';
-
-const mock_best_products: ThumbnailGoods[] = [
-  {
-    id: 1,
-    thumbnailUrl: mockProductImagePath,
-    title: '맥쥬짠',
-    price: 10000,
-    isGreen: true,
-    isBest: true,
-    discountRate: 0,
-  },
-  { id: 2, thumbnailUrl: mockProductImagePath, title: '맥쥬짠', price: 10000, isNew: true, discountRate: 0 },
-  { id: 3, title: 'NoImage 맥쥬짠', price: 10000, isSale: true, discountRate: 20 },
-  { id: 4, thumbnailUrl: mockProductImagePath, title: '맥쥬짠', price: 10000, discountRate: 0 },
-];
+import SideBar from '@src/pages/Main/SideBar/SideBar';
 
 const Main = () => {
   const [promotions, setPromotions] = useState<Promotion[]>([]);
@@ -61,7 +44,7 @@ const Main = () => {
           <GoodsSection sectionTitle='새로 나왔어요' goodsList={mainGoodsListMap.discountGoodsList} itemBoxSize='big' />
           <GoodsSection sectionTitle='지금 할인 중' goodsList={mainGoodsListMap.latestGoodsList} itemBoxSize='big' />
         </MainContentContainer>
-        <SideBar goodsList={mock_best_products} />
+        <SideBar />
         <FooterContainer>
           <Footer />
         </FooterContainer>
